feat: replace /api/stats with /api/views, IP-based 30min rate limit
